requirements.txt
Django==2.1.2
django-money==0.14.3
djangorestframework==3.8.2
